Add files via upload
append-sort.js was the file that got me through the first test set of the append-sort challenge, but failed in the second set.  It didn't properly handle sequences like (110, 1) => (110, 111).  After the challenge ended, I conjured better-append.js which took a smarter approach to the process, and succeeded in practice mode.